+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-## [0.1.0] - 2021-10-27
+- `operator-rs` `0.3.0` → `0.4.0` ([#20]).
+- Adapted pod image and container command to docker image ([#20]).
+- Adapted documentation to represent new workflow with docker images ([#20]). 
 
+[#20]: https://github.com/stackabletech/hdfs-operator/pull/20
+
+## [0.1.0] - 2021-10-27
 
 ### Changed
 - Switched to operator-rs tag 0.3.0 ([#13])